chore: apply @mini-max changes for PR #8866
