+++ src/refactor/src/codemods/index.ts
@@ -1,3 +1,4 @@
+export { applySourceTextEdits } from "./codemod-helpers.js";
 export * as GlobalvarToGlobal from "./globalvar-to-global/index.js";
 export * as LoopLengthHoisting from "./loop-length-hoisting/index.js";
 export * as NamingConvention from "./naming-convention/index.js";
