+++ src/refactor/src/codemods/loop-length-hoisting/loop-length-hoisting-codemod.ts
@@ -1,6 +1,7 @@
 import { Core } from "@gmloop/core";
 import { Parser } from "@gmloop/parser";
 
+import { applySourceTextEdits } from "../codemod-helpers.js";
 import type {
     LoopLengthHoistingCodemodOptions,
     LoopLengthHoistingCodemodResult,
@@ -187,32 +188,6 @@ function createLoopLengthHoistRewrite(parameters: {
         callRewrites,
         reportOffset: firstCall.callStart
     });
-}
-
-/**
- * Apply a list of non-overlapping edits to `sourceText` using a left-to-right
- * string builder.  Edits are sorted in **ascending** order so the result is
- * assembled in a single forward pass without intermediate string copies,
- * matching the same pattern used by `applyGroupedTextEditsToContent` in the
- * refactor engine.
- */
-function applySourceTextEdits(sourceText: string, edits: ReadonlyArray<LoopLengthHoistingEdit>): string {
-    if (edits.length === 0) {
-        return sourceText;
-    }
-
-    const sorted = [...edits].toSorted((left, right) => left.start - right.start || left.end - right.end);
-    let result = "";
-    let cursor = 0;
-
-    for (const edit of sorted) {
-        result += sourceText.slice(cursor, edit.start);
-        result += edit.text;
-        cursor = edit.end;
-    }
-
-    result += sourceText.slice(cursor);
-    return result;
 }
 
 /**
