+++ src/refactor/test/codemod-execution-performance.test.ts
@@ -34,12 +34,11 @@ import assert from "node:assert/strict";
 import test from "node:test";
 
 import { Refactor } from "../index.js";
-import type {
-    ConfiguredCodemodRunResult,
-    NamingConventionTarget,
-    PartialSemanticAnalyzer,
-    RefactorProjectConfig
-} from "../src/types.js";
+import type { ConfiguredCodemodRunResult, NamingConventionTarget, RefactorProjectConfig } from "../src/types.js";
+import {
+    buildNamingConventionSemanticStub,
+    createSyntheticLocalNamingFixture
+} from "./test-helpers/naming-convention-performance.js";
 import { measureMedianDurationMs } from "./test-helpers/performance-timing.js";
 
 const FILE_COUNT = 500;
@@ -59,81 +58,6 @@ const TOTAL_EDITS = FILE_COUNT * TARGETS_PER_FILE * 2;
  */
 const PERFORMANCE_THRESHOLD_MS = 1100;
 
-type SyntheticFileFixture = {
-    sourceText: string;
-    targets: Array<NamingConventionTarget>;
-};
-
-function createSyntheticFixture(filePath: string, fileIndex: number, targetsPerFile: number): SyntheticFileFixture {
-    const lines: Array<string> = [];
-    const targets: Array<NamingConventionTarget> = [];
-    let offset = 0;
-
-    for (let targetIndex = 0; targetIndex < targetsPerFile; targetIndex += 1) {
-        const currentName = `bad_name_${fileIndex}_${targetIndex}`;
-        const declarationLine = `var ${currentName} = ${targetIndex};\n`;
-        const referenceLine = `show_debug_message(${currentName});\n`;
-        const declarationStart = offset + declarationLine.indexOf(currentName);
-        const referenceStart = offset + declarationLine.length + referenceLine.indexOf(currentName);
-
-        lines.push(declarationLine, referenceLine);
-        targets.push({
-            name: currentName,
-            category: "localVariable",
-            path: filePath,
-            scopeId: `scope:${fileIndex}:${targetIndex}`,
-            symbolId: null,
-            occurrences: [
-                {
-                    path: filePath,
-                    start: declarationStart,
-                    end: declarationStart + currentName.length,
-                    kind: Refactor.OccurrenceKind.DEFINITION,
-                    scopeId: `scope:${fileIndex}:${targetIndex}`
-                },
-                {
-                    path: filePath,
-                    start: referenceStart,
-                    end: referenceStart + currentName.length,
-                    kind: Refactor.OccurrenceKind.REFERENCE,
-                    scopeId: `scope:${fileIndex}:${targetIndex}`
-                }
-            ]
-        });
-
-        offset += declarationLine.length + referenceLine.length;
-    }
-
-    return {
-        sourceText: lines.join(""),
-        targets
-    };
-}
-
-function buildSemanticStub(targetsByFile: Map<string, Array<NamingConventionTarget>>): PartialSemanticAnalyzer {
-    return {
-        listNamingConventionTargets: async (filePaths?: Array<string>) => {
-            const selectedPaths = filePaths === undefined ? null : new Set(filePaths);
-            const matchingTargets: Array<NamingConventionTarget> = [];
-
-            for (const [filePath, targets] of targetsByFile.entries()) {
-                const resourcePath = filePath.replace(/\.gml$/i, ".yy");
-                if (selectedPaths !== null && !selectedPaths.has(filePath) && !selectedPaths.has(resourcePath)) {
-                    continue;
-                }
-
-                matchingTargets.push(...targets);
-            }
-
-            return matchingTargets;
-        },
-        validateEdits: async () => ({
-            errors: [],
-            warnings: []
-        })
-    };
-}
-
 void test(`end-to-end codemod execution stays within regression threshold (${FILE_COUNT} files × ${TARGETS_PER_FILE} targets = ${TOTAL_EDITS} edits)`, async () => {
     const projectRoot = "/project";
     const sourceTexts = new Map<string, string>();
@@ -141,12 +65,12 @@ void test(`end-to-end codemod execution stays within regression threshold (${FIL
     const gmlFilePaths = Array.from({ length: FILE_COUNT }, (_, fileIndex) => `scripts/script_${fileIndex}.gml`);
 
     for (const [fileIndex, filePath] of gmlFilePaths.entries()) {
-        const fixture = createSyntheticFixture(filePath, fileIndex, TARGETS_PER_FILE);
+        const fixture = createSyntheticLocalNamingFixture(filePath, fileIndex, TARGETS_PER_FILE);
         sourceTexts.set(filePath, fixture.sourceText);
         targetsByFile.set(filePath, fixture.targets);
     }
 
-    const semantic = buildSemanticStub(targetsByFile);
+    const semantic = buildNamingConventionSemanticStub(targetsByFile);
     const engine = new Refactor.RefactorEngine({ semantic });
 
     const config: RefactorProjectConfig = {
