+++ src/refactor/test/naming-convention-performance.test.ts
@@ -2,143 +2,17 @@ import assert from "node:assert/strict";
 import test from "node:test";
 
 import { Refactor } from "../index.js";
-import type {
-    ConfiguredCodemodRunResult,
-    NamingConventionTarget,
-    PartialSemanticAnalyzer,
-    RefactorProjectConfig
-} from "../src/types.js";
+import type { NamingConventionTarget, PartialSemanticAnalyzer } from "../src/types.js";
+import {
+    buildNamingConventionCodemodExecutor,
+    buildNamingConventionSemanticStub,
+    createSyntheticLocalNamingFixture
+} from "./test-helpers/naming-convention-performance.js";
 import { measureMedianDurationMs } from "./test-helpers/performance-timing.js";
 
 const FILE_COUNT = 180;
 const TARGETS_PER_FILE = 32;
 const PERFORMANCE_THRESHOLD_MS = 700;
-const DEFAULT_LOCAL_VARIABLE_NAMING_CONFIG: RefactorProjectConfig["codemods"]["namingConvention"] = {
-    rules: {
-        localVariable: {
-            caseStyle: "camel"
-        }
-    }
-};
-
-type SyntheticFileFixture = {
-    sourceText: string;
-    targets: Array<NamingConventionTarget>;
-};
-
-function createSyntheticLocalNamingFixture(
-    filePath: string,
-    fileIndex: number,
-    targetsPerFile: number
-): SyntheticFileFixture {
-    const lines: Array<string> = [];
-    const targets: Array<NamingConventionTarget> = [];
-    let offset = 0;
-
-    for (let targetIndex = 0; targetIndex < targetsPerFile; targetIndex += 1) {
-        const currentName = `bad_name_${fileIndex}_${targetIndex}`;
-        const declarationLine = `var ${currentName} = ${targetIndex};\n`;
-        const referenceLine = `show_debug_message(${currentName});\n`;
-        const declarationStart = offset + declarationLine.indexOf(currentName);
-        const referenceStart = offset + declarationLine.length + referenceLine.indexOf(currentName);
-
-        lines.push(declarationLine, referenceLine);
-        targets.push({
-            name: currentName,
-            category: "localVariable",
-            path: filePath,
-            scopeId: `scope:${fileIndex}:${targetIndex}`,
-            symbolId: null,
-            occurrences: [
-                {
-                    path: filePath,
-                    start: declarationStart,
-                    end: declarationStart + currentName.length,
-                    kind: Refactor.OccurrenceKind.DEFINITION,
-                    scopeId: `scope:${fileIndex}:${targetIndex}`
-                },
-                {
-                    path: filePath,
-                    start: referenceStart,
-                    end: referenceStart + currentName.length,
-                    kind: Refactor.OccurrenceKind.REFERENCE,
-                    scopeId: `scope:${fileIndex}:${targetIndex}`
-                }
-            ]
-        });
-
-        offset += declarationLine.length + referenceLine.length;
-    }
-
-    return {
-        sourceText: lines.join(""),
-        targets
-    };
-}
-
-/**
- * Build a minimal {@link PartialSemanticAnalyzer} stub that returns pre-built
- * naming targets for the given file-to-targets map.  Accepts an optional
- * callback that is invoked on every call, allowing callers to track
- * invocation counts or perform side effects per query.
- */
-function buildNamingConventionSemanticStub(
-    targetsByFile: Map<string, Array<NamingConventionTarget>>,
-    onCall?: () => void
-): PartialSemanticAnalyzer {
-    return {
-        listNamingConventionTargets: async (filePaths?: Array<string>) => {
-            onCall?.();
-            const selectedPaths = filePaths === undefined ? null : new Set(filePaths);
-            const matchingTargets: Array<NamingConventionTarget> = [];
-
-            for (const [filePath, targets] of targetsByFile.entries()) {
-                const resourcePath = filePath.replace(/\.gml$/i, ".yy");
-                if (selectedPaths !== null && !selectedPaths.has(filePath) && !selectedPaths.has(resourcePath)) {
-                    continue;
-                }
-
-                matchingTargets.push(...targets);
-            }
-
-            return matchingTargets;
-        },
-        validateEdits: async () => ({
-            errors: [],
-            warnings: []
-        })
-    };
-}
-
-/**
- * Build the {@link Refactor.RefactorEngine.executeConfiguredCodemods} executor
- * used by both stress tests.  Each test supplies its own engine, file list, and
- * source-text map so the captured closure variables differ, while the call-site
- * shape is shared through this factory.
- */
-function buildNamingConventionCodemodExecutor(
-    engine: InstanceType<typeof Refactor.RefactorEngine>,
-    gmlFilePaths: Array<string>,
-    sourceTexts: Map<string, string>,
-    projectRoot: string,
-    namingConfig?: RefactorProjectConfig["codemods"]["namingConvention"]
-): () => Promise<ConfiguredCodemodRunResult> {
-    const config: RefactorProjectConfig = {
-        codemods: {
-            namingConvention: namingConfig ?? DEFAULT_LOCAL_VARIABLE_NAMING_CONFIG
-        }
-    };
-
-    return () =>
-        engine.executeConfiguredCodemods({
-            projectRoot,
-            targetPaths: [projectRoot],
-            gmlFilePaths,
-            config,
-            readFile: async (filePath) => sourceTexts.get(filePath) ?? "",
-            dryRun: true
-        });
-}
 
 void test("namingConvention stress test stays within the selected-file planning threshold", async () => {
     const projectRoot = "/project";
